Added user level suggestion list and fixed UI issue's

+++ anuvaad-fe/anuvaad-webapp/src/ui/containers/web/UserGlossary/ViewUserGlossary.jsx
@@ -75,14 +75,18 @@ class MyGlossary extends React.Component {
         }
     }
 
-    makeDeleteGlossaryAPICall = (userId, src, tgt, locale, reverseLocale, context, bulkDelete = false, deletionArray = []) => {
+    makeDeleteGlossaryAPICall = (userId, orgID, src, tgt, locale, reverseLocale, context, bulkDelete = false, deletionArray = []) => {
+        console.log(`userId, src, tgt, locale, reverseLocale, context, bulkDelete = false, deletionArray = []`);
+        console.log(userId, src, tgt, locale, reverseLocale, context, bulkDelete, deletionArray);
         this.setState({ open: true, message: 'Glossary deletion in progress...', variant: 'info' })
-        let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
-        let apiObj = new DeleteGlossary(userId, src, tgt, locale, reverseLocale, context, bulkDelete, deletionArray)
+        // let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+        let apiObj = new DeleteGlossary(userId, orgID, src, tgt, locale, reverseLocale, context, bulkDelete, deletionArray);
+        let payloadObj =  apiObj.getBody();
+        payloadObj.orgID && delete payloadObj.orgID;
         fetch(apiObj.apiEndPoint(), {
             method: 'post',
             headers: apiObj.getHeaders().headers,
-            body: JSON.stringify(apiObj.getBody())
+            body: JSON.stringify(payloadObj)
         })
             .then(async res => {
                 if (res.ok) {
@@ -168,7 +172,8 @@ class MyGlossary extends React.Component {
 
     handleDeleteGlossary = (dataArray) => {
         let reverseLocale = dataArray[3].split("|").reverse().join("|");
-        this.makeDeleteGlossaryAPICall(dataArray[2], dataArray[0], dataArray[1], dataArray[3], reverseLocale, dataArray[4])
+        let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+        this.makeDeleteGlossaryAPICall(dataArray[2], orgID, dataArray[0], dataArray[1], dataArray[3], reverseLocale, dataArray[4])
     }
 
     handleClose = () => {
@@ -179,8 +184,9 @@ class MyGlossary extends React.Component {
         let isOrg = delete_glossary.isOrg(this.props.glossaryData, this.state.rowsToDelete)
         if (!isOrg) {
             let userId = JSON.parse(localStorage.getItem("userProfile")).userID
-            let rowsToBeDeleted = delete_glossary.getBulkDeletionArray(this.props.glossaryData, this.state.rowsToDelete)
-            this.makeDeleteGlossaryAPICall(userId, "", "", "", "", "JUDICIARY", true, rowsToBeDeleted)
+            let rowsToBeDeleted = delete_glossary.getBulkDeletionArray(this.props.glossaryData, this.state.rowsToDelete);
+            let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+            this.makeDeleteGlossaryAPICall(userId, orgID, "", "", "", "", "JUDICIARY", true, rowsToBeDeleted)
         } else {
             this.setState({ open: true, message: "Cannot delete glossary of type Organization..", variant: "error" })
         }
